sticks = floats?

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Robot/opmode/teleop/competitionOpMode.java
@@ -16,15 +16,15 @@ public class competitionOpMode extends LinearOpMode {
         while (opModeIsActive()) {
 
 
-            double y = -gamepad1.left_stick_y; // Remember, Y stick is reversed!
-            double x = gamepad1.left_stick_x;
-            double rx = gamepad1.right_stick_x;
+            float y = -gamepad1.left_stick_y; // Remember, Y stick is reversed!
+            float x = gamepad1.left_stick_x;
+            float rx = gamepad1.right_stick_x;
 
-            double denominator = Math.max(Math.abs(y) + Math.abs(x) + Math.abs(rx), 1);
-            double frontLeftPower = (y + x + rx) / denominator;
-            double backLeftPower = (y - x + rx) / denominator;
-            double frontRightPower = (y - x - rx) / denominator;
-            double backRightPower = (y + x - rx) / denominator;
+            float denominator = Math.max(Math.abs(y) + Math.abs(x) + Math.abs(rx), 1);
+            float frontLeftPower = (y + x + rx) / denominator;
+            float backLeftPower = (y - x + rx) / denominator;
+            float frontRightPower = (y - x - rx) / denominator;
+            float backRightPower = (y + x - rx) / denominator;
 
             Drivetrain.drive(frontLeftPower, frontRightPower, backLeftPower, backRightPower);
